chore: Organized type files by domain
